backend/requirements.txt
fastapi
uvicorn[standard]
supabase
scikit-learn
pandas
numpy
python-dotenv
pydantic
